Added Kinect cursor (#7)
* Working Kinect Cursor!

It is currently quite jumpy, however it works! Due to the jumpiness it is likely we will have to massively increase the size of the buttons.

There are also some overlap issues between Table 4 and 5 during selection. I am not sure why, but I suspect the 'cursor' is wide enough to get into the hover sections of both.

* Reduced cursor size

As suspected it deals with overlap issues. Will do a better fix (ie reduce size of hover sections in svg) later.